Add files via upload
This simple calculator can be a useful tool for students, professionals, and anyone who needs to perform basic arithmetic operations quickly and easily. It can also be a good starting point for building more advanced calculators or other mathematical applications. Although it is a simple program, it demonstrates the power and flexibility of Python as a programming language.